+++ src/Command/GenerateCommand.php
@@ -2,7 +2,7 @@
 
 namespace SatisGen\Command;
 
-use SatisGen\Config\ReaderInterface;
+use SatisGen\Config\AdvancedConfigReaderInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -17,7 +17,7 @@ class GenerateCommand extends Command
     private $filesystem;
     private $config;
 
-    public function __construct(Filesystem $filesystem, ReaderInterface $config)
+    public function __construct(Filesystem $filesystem, AdvancedConfigReaderInterface $config)
     {
         $this->filesystem = $filesystem;
         $this->config = $config;
@@ -49,13 +49,12 @@ class GenerateCommand extends Command
         $inputFile  = $input->getArgument('input_file') ?: 'satis.php';
         $outputFile = $input->getArgument('output_file') ?: 'satis.json';
 
-        $helper = $this->getHelper('question');
-        $this->config->setInput($input);
-        $this->config->setOutput($output);
-        $this->config->setHelper($helper);
+        $this->config->getInputReader()->setInput($input);
+        $this->config->getInputReader()->setOutput($output);
+        $this->config->getInputReader()->setQuestionHelper($this->getHelper('question'));
 
         if ($filesystem->exists($inputFile)) {
-            $output->writeln('<info>Generating... </info>');
+            $output->write('<info>Generating... </info>');
             /* TODO: Fixed by https://github.com/symfony/symfony/pull/14580
             try {
                 $filesystem->dumpFile($outputFile, $this->generate($inputFile));
@@ -76,6 +75,5 @@ class GenerateCommand extends Command
         ob_start();
         require $file;
         return ob_get_clean();
-
     }
 }