+++ src/Config/AdvancedConfigReaderInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace SatisGen\Config;
+
+interface AdvancedConfigReaderInterface {
+
+    public function getConfig($name, $description, $default = null, $validator = null, $secure = false);
+
+}