+++ src/Config/ConfigReaderInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace SatisGen\Config;
+
+interface ConfigReaderInterface {
+
+    public function getConfig($name, $default = null);
+
+}