+++ src/Config/ConfigWriterInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace SatisGen\Config;
+
+interface ConfigWriterInterface {
+
+    public function setConfigs($configs);
+
+    public function setConfig($name, $value);
+
+}