+++ src/Config/ReaderInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace SatisGen\Config;
-
-interface ReaderInterface {
-
-    public function getEnv($name, $description, $filter = FILTER_DEFAULT, $options = null, $secure = false);
-
-}